requirements.txt
azure==0.8.4
Flask==0.10.1
Flask-MySQLdb==0.2.0
Jinja2==2.8
MarkupSafe==0.23
Reindent==0.1.1
Werkzeug==0.9.4
argparse==1.2.1
gunicorn==18.0
itsdangerous==0.24
